Рефакторинг слайдера, favicon
