--- file.json
@@ -1,76 +1,0 @@
-[
-  {
-    "id": 1,
-    "title": "Мужские Кроссовки Nike Blazer Mid Suede",
-    "img": "./img/sneakers/1.jpg",
-    "price": 12999,
-    "sale": 0.4
-  },
-  {
-    "id": 2,
-    "title": "Мужские Кроссовки Nike Air Max 270",
-    "img": "./img/sneakers/2.jpg",
-    "price": 12999
-  },
-  {
-    "id": 3,
-    "title": "Мужские Кроссовки Nike Blazer Mid Suede",
-    "img": "./img/sneakers/3.jpg",
-    "price": 8499,
-    "sale": 0.7
-  },
-  {
-    "id": 4,
-    "title": "Кроссовки Puma X Aka Boku Future Rider",
-    "img": "./img/sneakers/4.jpg",
-    "price": 8999
-  },
-  {
-    "id": 5,
-    "title": "Мужские Кроссовки Under Armour Curry 8",
-    "img": "./img/sneakers/5.jpg",
-    "price": 15199
-  },
-  {
-    "id": 6,
-    "title": "Мужские Кроссовки Nike Kyrie 7",
-    "img": "./img/sneakers/6.jpg",
-    "price": 11299
-  },
-  {
-    "id": 7,
-    "title": "Мужские Кроссовки Jordan Air Jordan 11",
-    "img": "./img/sneakers/7.jpg",
-    "price": 10799
-  },
-  {
-    "id": 8,
-    "title": "Мужские Кроссовки Nike LeBron XVIII",
-    "img": "./img/sneakers/8.jpg",
-    "price": 16499
-  },
-  {
-    "id": 9,
-    "title": "Мужские Кроссовки Nike Lebron XVIII Low",
-    "img": "./img/sneakers/9.jpg",
-    "price": 13999
-  },
-  {
-    "id": 10,
-    "title": "Мужские Кроссовки Nike Blazer Mid Suede",
-    "img": "./img/sneakers/10.jpg",
-    "price": 8499
-  },
-  {
-    "id": 11,
-    "title": "Кроссовки Puma X Aka Boku Future Rider",
-    "img": "./img/sneakers/11.jpg",
-    "price": 8999
-  },
-  {
-    "id": 12,
-    "title": "Мужские Кроссовки Nike Kyrie Flytrap IV",
-    "img": "./img/sneakers/12.jpg",
-    "price": 11299
-  }
-]
