+++ src/UI/Slider/_Slider.scss
